[#5/handle-permission] Admin can toggle ROLE_ADMIN of users

+++ src/Controller/UserController.php
@@ -3,10 +3,12 @@
 namespace App\Controller;
 
 use App\Entity\User;
-use App\Form\RegistrationFormType;
 use App\Form\UserType;
+use App\Handler\Forms\UserFormHandler;
 use App\Repository\UserRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
+use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -35,7 +37,7 @@ class UserController extends AbstractController
     }
 
     /**
-     * @Route("/users", name="user_index")
+     * @Route("/admin/users", name="user_index")
      * @param UserRepository $userRepository
      * @return \Symfony\Component\HttpFoundation\Response
      */
@@ -45,40 +47,55 @@ class UserController extends AbstractController
     }
 
     /**
-     * @Route("/users/{id}/edit", name="user_edit")
-     * @param User $user
+     * @Route("/admin/users/new", name="user_new")
      * @param Request $request
+     * @param UserFormHandler $formHandler
      * @return \Symfony\Component\HttpFoundation\RedirectResponse|\Symfony\Component\HttpFoundation\Response
      */
-    public function edit(User $user, Request $request)
+    public function new(Request $request, UserFormHandler $formHandler)
     {
-        $form = $this->createForm(RegistrationFormType::class, $user);
-
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $password = $this->passwordEncoder->encodePassword($user, $user->getPassword());
-            $user->setPassword($password);
-
-            $this->entityManager->flush();
-
-            $this->addFlash('success', "L'utilisateur a bien été modifié");
-
+        $user = new User;
+        if ($formHandler->handle($request, $user,UserType::class)) {
+            $this->addFlash('success', "L'utilisateur a bien été ajouté");
             return $this->redirectToRoute('user_index');
         }
 
-        return $this->render('user/edit.html.twig', ['form' => $form->createView(), 'user' => $user]);
+        return $this->render('user/new.html.twig', ['form' => $formHandler->createView(), 'user' => $user]);
+    }
+
+    /**
+     * @Route("/users/{id}/edit", name="user_edit")
+     * @param User $user
+     * @param Request $request
+     * @param UserFormHandler $formHandler
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse|\Symfony\Component\HttpFoundation\Response
+     */
+    public function edit(User $user, Request $request, UserFormHandler $formHandler)
+    {
+        $this->denyAccessUnlessGranted('USER_EDIT', $user);
+
+        if ($formHandler->handle($request, $user,UserType::class)) {
+            $this->addFlash('success', "L'utilisateur a bien été modifié");
+            return $this->redirectToRoute('user_index');
+        }
+
+        return $this->render('user/edit.html.twig', ['form' => $formHandler->createView(), 'user' => $user]);
     }
 
     /**
      * @Route("/users/{id}/delete", name="user_delete")
      * @param User $user
      * @param Request $request
+     * @param SessionInterface $session
      * @return \Symfony\Component\HttpFoundation\RedirectResponse
      */
-    public function delete(User $user, Request $request)
+    public function delete(User $user, Request $request, SessionInterface $session)
     {
+        $this->denyAccessUnlessGranted('USER_DELETE', $user);
         if ($this->isCsrfTokenValid("delete" . $user->getId(), $request->get("_token"))) {
+            if ($this->getUser()->getId() === $user->getId()) {
+                $session->invalidate();
+            }
             $this->entityManager->remove($user);
             $this->entityManager->flush();
 
@@ -86,6 +103,6 @@ class UserController extends AbstractController
             return $this->redirectToRoute('user_index');
         }
         $this->addFlash('error', 'Token invalide');
-        return $this->redirectToRoute('user_index', [], 403);
+        return $this->redirectToRoute('user_index');
     }
 }
